requirements.txt
psycopg2-binary==2.9.9
python-dotenv==0.9.9
argon2-cffi==25.1.10
sqlmodel==0.0.25